fix API upload and API getHistory

+++ backend/image-optimizer/src/main/java/com/aws/image_optimizer/entity/ImageMetadata.java
@@ -68,7 +68,7 @@ public class ImageMetadata {
     private Double compressionRatio;
 
 
-    private Long processingTimeMs;
+    private Double processingTimeMs;
